getLocation
