+++ server/models/User.js
@@ -1,0 +1,14 @@
+const mongoose = require('mongoose')
+const Schema = mongoose.Schema
+
+const userSchema = new Schema({
+    firstName: String,
+    lastName: String,
+    age: Number,
+    location: {type:Schema.Types.ObjectId,ref:"location"},
+    trips: [{type:Schema.Types.ObjectId,ref:"trip"}]
+})
+
+const User = mongoose.model("user",userSchema)
+
+module.exports = User
